move pieces into an object
So we can do lookup on them instead of doing heavy switching to
find them, or creating new pieces.

This is part of an early decision to let there be exactly *one*
piece object of each type. This might be the most controversial
design decision in the entire game, but it works because pieces
don't have any mutable state, only some properties it got at
creating time and some methods.

The advantage is that with exactly one piece of each type, we
can compare using == or === equality, which will compare two
pieces using reference equality. If there was the possibility in
the model of having two distinct objects representing the same
piece type, then some kind of .equals method would have to be
introduced to check whether two pieces are of the same type.

This particular change is going in because promotion requires
us to do lookup on pieces, going "give me the piece of the
same color, but of the given type". This is much easier to do
by placing the pieces in an object that we can then do lookup
on.